src/requirements.txt
git+https://github.com/zimmerman-zimmerman/fulltext.git
django-nested-admin==2.2.6
requests==2.12.4
Django==1.11
django-cors-headers==1.2.2
django-debug-toolbar==1.6
django-filter==1.0.2
djangorestframework==3.5.3
factory-boy==2.8.1
lxml==3.7.1
psycopg2==2.7.4
pysqlite==2.8.3
pytest==3.0.5
pytest-django==3.1.2
python-coveralls==2.9.0
ujson==1.35
xlrd==1.0.0
django-rest-swagger==2.1.2
django-rest-framework-docs==0.1.7
django-rq==0.9.5
rq==0.7.1
textract==1.6.1
rq-scheduler==0.7.0
django-supervisor==0.4.0
supervisor==3.3.1
django-extensions==1.7.8
django-rest-auth==0.9.1
django-allauth==0.31.0
rfc6266==0.0.4
python-dateutil==2.6.0
redis==2.10.5
django-redis-cache==1.7.1
tweepy==3.5.0
uwsgi==2.0.15
zeep==2.5.0
nltk==3.2.5
Unidecode==1.0.22
django-after-response==0.2.2
pdf2image==1.0.0
